starting test_schemas migration
