+++ src/snowflakes/tests/test_schemas.py
@@ -7,48 +7,51 @@ SCHEMA_FILES = [
 ]
 
 
-@pytest.mark.parametrize('schema', SCHEMA_FILES)
-def test_load_schema(schema):
-    from snovault.schema_utils import load_schema
-    assert load_schema('snowflakes:schemas/%s' % schema)
+# @pytest.mark.parametrize('schema', SCHEMA_FILES)
+# def test_load_schema(schema):
+#     from snovault.schema_utils import load_schema
+#     assert load_schema('snowflakes:schemas/%s' % schema)
 
-
+# XXX: Can't migrate?
 def test_linkTo_saves_uuid(root, submitter, lab):
     item = root['users'][submitter['uuid']]
     assert item.properties['submits_for'] == [lab['uuid']]
 
 
+# XXX: add schema that uses mixin properties to 
+# testing-views.py in snovault
 def test_mixinProperties():
     from snovault.schema_utils import load_schema
     schema = load_schema('snowflakes:schemas/access_key.json')
     assert schema['properties']['uuid']['type'] == 'string'
 
 
-def test_dependencies(testapp):
-    collection_url = '/testing-dependencies/'
-    testapp.post_json(collection_url, {'dep1': 'dep1', 'dep2': 'dep2'}, status=201)
-    testapp.post_json(collection_url, {'dep1': 'dep1'}, status=422)
-    testapp.post_json(collection_url, {'dep2': 'dep2'}, status=422)
-    testapp.post_json(collection_url, {'dep1': 'dep1', 'dep2': 'disallowed'}, status=422)
+# def test_dependencies(testapp):
+#     collection_url = '/testing-dependencies/'
+#     testapp.post_json(collection_url, {'dep1': 'dep1', 'dep2': 'dep2'}, status=201)
+#     testapp.post_json(collection_url, {'dep1': 'dep1'}, status=422)
+#     testapp.post_json(collection_url, {'dep2': 'dep2'}, status=422)
+#     testapp.post_json(collection_url, {'dep1': 'dep1', 'dep2': 'disallowed'}, status=422)
 
 
+# XXX: none of the test schemas utilize not_collection_default_page
 def test_page_schema_validates_parent_is_not_collection_default_page(testapp):
     res = testapp.post_json('/pages/', {'name': 'users', 'title': 'Users'})
     uuid = res.json['@graph'][0]['@id']
     testapp.post_json('/pages/', {'parent': uuid, 'name': 'test', 'title': 'Test'}, status=422)
 
 
-def test_changelogs(testapp, registry):
-    from snovault import TYPES
-    for typeinfo in registry[TYPES].by_item_type.values():
-        changelog = typeinfo.schema.get('changelog')
-        if changelog is not None:
-            res = testapp.get(changelog)
-            assert res.status_int == 200, changelog
-            assert res.content_type == 'text/markdown'
+# def test_changelogs(testapp, registry):
+#     from snovault import TYPES
+#     for typeinfo in registry[TYPES].by_item_type.values():
+#         changelog = typeinfo.schema.get('changelog')
+#         if changelog is not None:
+#             res = testapp.get(changelog)
+#             assert res.status_int == 200, changelog
+#             assert res.content_type == 'text/markdown'
 
 
-def test_schemas_etag(testapp):
-    etag = testapp.get('/profiles/', status=200).etag
-    assert etag
-    testapp.get('/profiles/', headers={'If-None-Match': etag}, status=304)
+# def test_schemas_etag(testapp):
+#     etag = testapp.get('/profiles/', status=200).etag
+#     assert etag
+#     testapp.get('/profiles/', headers={'If-None-Match': etag}, status=304)
